Separate nova live migration to nova/migration network role
This should work, because Nova submit to *live_migration_uri*
host name, which compute node registered in the Conductor.
If we make additional nostnames with more specific IP addresses
and make this resolvable.

cfg.StrOpt('live_migration_uri',
           default="qemu+tcp://%s/system",
           help='Migration target URI '
                '(any included "%s" is replaced with '
                'the migration target hostname)'),
that it will substitute %s with hostname and we get prefixed resolvable by hostname url.

Blueprint: templates-for-networking
Change-Id: I3e4654ff74382cd66136c1758f055ef4ac4f2109