Add second Sahara endpoint
When using the keystone static catalog file to register endpoints,
an endpoint registered (correctly) as catalog.region.data_processing
gets read as "data-processing" by keystone.
Thus, when Sahara looks for an endpoint, it is unable to find one
for data_processing.

This causes a problem with the commandline interface and the dashboard.

Change-Id: Id15a1fcb403d8cc8db4f485d1ef7c6e94ed5bcc1
Closes-bug: #1431983