Install proper mysql-client package
We should not install default mysql-client package, that is
defined in mysql::params::client_package_name. But instead we need
to calculate proper version of mysql-client in exactly the same
way as we do it in database.pp task when evaluate mysql::server
class. To do so we introduce new class galera::client and declare
it instead of ::mysql.

Change-Id: Ib118450656f99f68358e84a61b1684b8025e483c
Closes-bug: #1476599