Fix keystone provider to work with v3 auth_url
Fuel's current openrc uses v2.0 keystone API, but
the keystone providers use v3. It is necessary to
strip the version string from OS_AUTH_URL in order
for them to function. This was not a bug on hosts
where keystone is on the same host as where the
resources are created because keystone.conf is used.

Summary of changes:
* Added auth_url_version function to openstacklib credentials.rb
* Fetch auth_url from credentials class in keystone base provider
* Added OS_PROJECT_NAME to openrc (for openrc validation in openstacklib)
* Moved auth_file to be evaluated after keystone admin role
* Added openrc purge before evaluating keystone admin role
* Add workaround for missing HOME for finding /root/openrc
* Add spec tests for openstacklib and keystone

Upstream-Change-Id: I63278fbb34d4a2e37861d0ec411226feb27ef9dc
Upstream-Change-Id: I33c932f3dd4bacf95e44d5291e7b3bdb1597974f
Change-Id: Ifd0f378d189424e8fdf65f13a0f95fe8f50d4a45
Closes-Bug: #1479382