Fix keystone provider to work with v3 auth_url
Fuel's current openrc uses v2.0 keystone API, but
the keystone providers use v3. It is necessary to
set the OS_AUTH_URL to v3 in order for them to
function. This was not a bug on hosts where keystone
is on the same host as where the resources are
created.

Summary of changes:
* Added auth_url_version function to openstacklib credentials.rb
* Added call to auth_url_version in keystone base provider
* Added OS_PROJECT_NAME to openrc (for openrc validation in keystone)

Change-Id: Ifd0f378d189424e8fdf65f13a0f95fe8f50d4a45
Closes-Bug: #1479879
Closes-Bug: #1479382