Change rabbitmq internal api call to public one
The rabbit_auth_backend_internal:check_user_login is an internal api
call which was removed in Rabbit 3.5.x. Instead we should use
rabbit_access_control:check_user_pass_login which is public and it works
in both 3.3.5 we have now and 3.5.4 we are going to migrate to later.

For further details see
https://github.com/puppetlabs/puppetlabs-rabbitmq/pull/332

Change-Id: I6f11e5d4d832c703641b833480092c342c3dc71d