Parametrize Ceph service title by OS variant
Currently, the title and command for starting/restarting service in Ceph module
is incorrect for Ubuntu system (should be 'ceph-all', not just 'ceph'). Because
of this, on the reinstalled system with existing Ceph OSD devices the Puppet
module fails to activate those devices when the service is started.

Add parameter to params.pp manifest to specify a title of the service depending
on the operating system family.

Change-Id: I86f8b455479f84fd061ef4985a5e1371c039cf77
Closes-bug: 1474510