Set single memcache host for keystone cache
Now we have a list of memcache hosts in keystone.conf and each
host is a controller. When some of controllers goes down we need to
wait socket_timeout seconds for check this host. If we set a single
host to memcache_servers we don't need wait and connect somewhere.

Change-Id: I40523393e53cf2b458bc5ced33a6271d4869d450
Closes-bug: #1471264