Fix OCF stop errors for mysql
W/o this fix, the pacemaker mysql resource
would became unmanaged when failed to kill
mysql on OCF stop action. For mysql OCF,
this possible only when the kill command failed:
ESRCH
   No process or process group can be found
corresponding to that specified by pid.

This is an issue as pacemaker resources must
never turn unmanaged for such cases (wrong pid)
and the stop action should not return error.
Instead we consider this case as "nothing to do
here, return OK" in the hope of the nearest start
action to fix this pid issue later.

The solution is:
- Do not exit with error from the stop() when
failed to terminate with SIGTERM. This impacts
nothing as there is the code below which
shutdowns the whole mysql process group anyway.
- Add a warning message instead
- Update the log message to provide all PIDs under
the process group being terminated

Closes-bug: #1388771

cherry-pick from 533372f2098e29b8d3bbded5308379331de2c803
Change-Id: I2c4ecc10e9129e94e5610d774c92f6eaf5228759
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>