Fix OCF stop errors for mysql
W/o this fix, the mysqld process would became
unmanaged in pacemaker when failed to process
the SIGTERM while stopping.

This is an issue as pacemaker resources must
never turn unmanaged and stop actions should
not return errors, unless there is a STONITH
configured in pacemaker.

The solution is:
- Do not exit with error from the stop() when
failed to terminate with SIGTERM. This impacts
nothing as there is the code below which
shutdowns the whole mysql process group anyway.
- Add a warning message instead
- Update the log message to provide all PIDs under
the process group being terminated

Closes-bug: #1388771

Change-Id: I2c4ecc10e9129e94e5610d774c92f6eaf5228759
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>