Redirect http horizon to https when https enabled
This change updates the haproxy configuration for horizon to redirect
http traffic to https when https is enabled to ensure that only https is
used. When ssl is not enabled, the horizon http endpoint is configured
as a regular horizon endpoint.

Change-Id: I2db1fce35d4686e746cc1753ceea5332ab1325cb
Closes-Bug: 1484580
DocImpact: Horizon via HTTP will only redirect to the HTTPS version if
SSL is enabled.