Use PublicURL as AdminURL endpoint as well
W/o this fix, administrative actions cannot
be performed from outside and bound to the
cloud management network only.

This is an issue as a user may want to automate
a cloud administration with external scripts. But
the management network VIP is currently put in
all of the admin endpoints and it must never be
exposed externally as there is too much sensitive
data flows inside.

The solution is to either to introduce a separate
OpenStack API Admin Network in Fuel, or to use the
PublicURL as the AdminURL as the default choice.
This patch implements the latter one and also allows
to override both Admin VIP and protocol via hiera.

DocImpact: Reference arch, Ops & Security guides -
administrative actions allowed to be executed via
the public endpoint using the same protocol as it was
configured for it (http or https).

Closes-bug: #1362641

Change-Id: I979fa63957d857dc664665a4faa606290b5d66b1
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>