Fix horizon http haproxy configuration
This change updates the server_names and ipaddresses for the http
horizon haproxy configuration when ssl is enabled. Currently when ssl is
enabled, a crit log message is thrown for "No content, source or symlink
specified" for the http horizon haproxy configuration because we are
passing empty arrays. We should be using 'undef' to prevent the
balancermembers from being configured.

Change-Id: Iada2ebbb61301e7d79dbcfc72d5a5ae07dd0052a
Closes-Bug: 1486076