Update haproxy restart mechanism
Previously, we were restarting haproxy for every balancermember
configuration even if there were no changes. This was resulting in
haproxy running multiple times. With this change we are creating a new
wrapper class called openstack::ha::haproxy_restart that contains a
single refreshonly exec called 'haproxy-restart'. This exec is notified
by new haproxy::listen and haproxy::balancermember configurations to
restart haproxy only if there is a change.  With this change, haproxy
will only restart if there is an actual configuration change and we will
only be executing a restart once.

Change-Id: I247b66c8b1cac04a1eeeb22377cd06809ef93636
Closes-Bug: 1485991