Fix missing OCF logging
W/o this patch, the most of the OCF agents
sending their own logs to nothere.

This is a major issue, because we want to
be able troubleshoot and keep OCF events
in syslog and collected at remote logs
server(s) as well.

The solution is to enable HA_LOGTAG and
HA_LOGFACILITY for each OCF script shipped
with the Fuel.

DocImpact: Ops guide. Logging events from
Fuel OCF agents will be collected both
locally in the /var/log/daemon.log and remotely,
with the file names based on the agent "foo-agent"
name as the "foo-agent-ocf.log". Note, that
RabbitMQ and Neutron agents OCF logs will
be kept as is due to backwards compatibility reasons.

Change-Id: Ida2d5fe31de42d31131e0f3a3e7b1a771caf8dfd
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>