Revert haproxy restart primitive back to reload
This commit switches back haproxy configuration reload
to use reload command of OCF script. Otherwise, for Ubuntu
since kernel 3.9 haproxy due to some race condition in
OCF script may start two times with different configs
as it uses REUSEPORT socket option. By using reload
we ensure that there is always one process on the listening
socket which has the latest config.

This fix is may be partial as we may also require additional work
on OCF script to introduce additional checks on pids and pidfiles.
It can also be tracked as a partial duplicate of the bug related
to better handling of processes manipulations in OCF scripts.

Change-Id: Iadad82d1733a3c22c470e1ddb761900304ee9753
Partial-bug: #1487900
Related-bug: #1425579
Related-bug: #1401850