Fixing dockerctl restore/stop
This change fixes the stop_container logic so that it correctly passes
the keys from the CONTAINER_NAMES array to the stop_container function
when 'all' is given. Previously the logic was incorrectly passing the
values from the CONTAINER_NAMES array so the docker stop command was
failing as part of a restore or a stop all.

Change-Id: Id1d879ce9e4439b521a807e8e4abe18aada420a3
Related-Bug: 1485985
Closes-Bug: 1488541