Ensure ceph properly configured for ephermal usage
This change ensures that ceph mon and the ceph compute configurations
are correctly applied when ceph is selected for only ephermal volumes.
Additionally, adding noop astute.yaml files for this type of deployment
where ephemeral_ceph is enabled since none of the previous files have
ephemeral_ceph set to true.

Change-Id: I9e86c1180d2317eab89babfd96a222bd8abce357
Co-Authored-By: Alex Schultz <aschultz@mirantis.com>
Closes-Bug: 1489532