Add module reset logic for librarian
This change adds the ability to do a reset or an update with the
updates_modules.sh script. The reset will do a git reset on the upstream
module if there are any manual changes. The update will do a librarian
update which will checkout the version specified in the Puppetfile.
These changes allow for better development testing as a developer can
test manual changes and version changes via the update_modules.sh
script. Additionally the fuel_noop_tests.rb script has been updated to
support the reset flag for developer testing.

Change-Id: Ib21b3ea210b554e1f58bb7a80453989ba8206104
Closes-Bug: 1489542