Add module reset logic for librarian
This change adds additional reset logic into the update_modules.sh
script to ensure that the contents of the module folder matches the
required version as specified in the Puppetfile. This change also
updates the fuel_noop_tests.rb script with the reset flag that can be
passed in to do the reset for testing. This reset flag is useful for
development, but should not be needed when run via jenkins as the
modules should be installed fresh for each run.

Change-Id: Ib21b3ea210b554e1f58bb7a80453989ba8206104
Closes-Bug: 1489542