Change protocol and hostname for novncproxy
To have an access to the VM console over HTTPS, we need to
give a right link to proxy via api.
Also, there is raises the need for user to setup DNS on
the computer from which he will try to access a VM console
accordingly to host name written in the certificate for Nova
VNC Proxy service.

Closes-Bug: #1489898
DocImpact

Change-Id: Ib27f3afa60a03800fb5da236dbb5379d759c2a07