Change protocol and hostname for novncproxy
In order to provide VM console access via HTTPS, we need to provide
the correct link to the proxy via the api.
Also, the cloud administrator will need to properly configure the DNS
for the clients to ensure they will be able to access the HTTPS
endpoint as the DNS name must match the certificate name for the Nova
VNC Proxy service.

Closes-Bug: #1489898
DocImpact

Change-Id: Ib27f3afa60a03800fb5da236dbb5379d759c2a07