DO NOT REVIEW: Get rid of version.yaml facts
/etc/version.yaml is built together with ISO and
it is not a part of any rpm package which contradicts
package based approach.

This patch introduces 2 other puppet facts which
are from the following files:
- /etc/fuel_release
- /etc/fuel_openstack_version
These files are to be packaged into RPM packages and built
using perestroika. Besides, it assumes there is is
PRODUCTION variable in astute.yaml which is used
to set $production (by default it is 'docker').

Depends-On: I02f32386232c94a0288877dfc78513a8bc74d23e
Depends-On: I6e867c4e118518ad02425b49bb067738d4da20c4
Depends-On: I175dc620cbdff3959519e3fc374894f93bba45ac
Implements blueprint: docker-images-perestroika

Change-Id: I5f264fbc8f80a2e92920484ef6a6e28e6c8e7e58