Add seed to fqdn_rotate
This change adds an additional seed to the fqdn_rotate for the apache
logrotation delay. We have seen that without the seed, some
configurations result in not enough randomness resulting in the same
delay for all controllers. By adding a seed string, the fqdn_rotate
command provides a better result.

Change-Id: I76f106d85ac363708e9fe35b04e56b189de1407f
Closes-Bug: 1491576