Move sysctl module to librarian
This commit removes the sysctl module from the fuel-library code and
replaces it to be managed via librarian. This change also includes an
update to version v0.0.11 as the previous version that was included did
not match an actual release. The upgrade also closes the incorrect noop
tests failures on OSX for Bug 1492438.

Partial blueprint: fuel-puppet-librarian

Change-Id: Ieeb39be2b05da7c7efeaf5a4909e32fc579d2cdf
Closes-Bug: 1492438