Moving rsyslog to librarian v2
This change moves the rsyslog module to librarian. This change also
updates the rsyslog module to v3.5.1 from our customized 1.1.4 version.
With the update, our usage of file resources to manage syslog
configuration files has also been updated to leverage the
rsyslog::snippet resource.

This change also cleans up the lint issues with the openstack::logging
class.

The previous version of this patch was reverted due to missed
configuration changes between how 1.1.4 managed rules and how 3.5.1
manages rules. The newer configurations use rulesets for the remote
rules and so the configuration file for the master's rsyslog needs to
target the remote ruleset for where to place remote files.

Partial blueprint: fuel-puppet-librarian

Change-Id: I345844d44545381f0d79b6113eb2dbcc515d77cb