(WIP) Dynamic dnsmasq config for admin network ranges
WORK IN PROGRESS

In order to successfully PXE/boot nodes from new racks we need to update
cobbler's dnsmasq configuration on changes in admin networks of any
nodegroup in any Fuel environment. Such changes should survive cobbler
container restarts and rebuilds.

In order to do so we need to execute the following on every change in
admin netwoks configuration of any nodegroup we have in Fuel:
- dump list of admin networks into /etc/hiera/networks.yaml file
  in mcollective container
- run 'puppet apply nailgin/examples/dhcp-ranges.pp' inside mcollective
  container to update config files in shared /etc/dnsmasq.d/ dir
- run 'cobbler sync' inside cobbler container to reload dnsmasq

Partial-bug: #1495593

Fuel-CI: disable
Change-Id: I9d7428c9fc21c705a1aee7fbca8003829a81e1d0