Fix openstack module tests
This change reworks the tests for the openstack module so they actually
function. This change includes basic tests for the
openstack::{logging,logrotate,controller,compute} classes. Additionally
some minor lint issues have been addressed to fix errors that were
preventing the syntax checker from passing.

This change is necessary so we can properly test changes when we need to
replace or upgrade some of the module dependencies for the openstack
module.

Partial blueprint: fuel-puppet-librarian

Change-Id: I4e2a1440fb805c3d8519fd242e71717c12cdb24a