Add cluster disk space monitor
This change adds a monitor into corosync/pacemaker to migrate services
if the monitored disks drop below 100M free.

Once the operator has resolved the full disk, they must clear the
alarm by running:

 crm node status-attr <hostname> delete "#health_disk"

After the alarm has been cleared, the services should be automatically
restarted.

This change is not a replacement for proper monitoring, but it will
properly shut down and migrate services if a controller runs out of disk
space.

DocImpact
Closes-Bug: 1493520

Change-Id: I8a2cb4bd8d0b6070400d13e25d2310f4777b9faf