Move horizon module to librarian
This commit removes 'horizon' module from fuel-library
and replaces it to be managed via librarian.
This commit also uses mos-8.0 branch of horizon module.
This branch is intermediate stable state of
current master and it contains one custom commit
https://review.fuel-infra.org/#/c/10938/
which should be reverted later (according to commit message).
Then it will be switched on stable/liberty.

Partial implements: blueprint fuel-puppet-librarian

Change-Id: I06efc29e8dbde2c8725e3b58f0f2b2b7e699bf9b