Update OCF stop actions to use procfs
This change updates the OCF scripts that currently kill a process to
check procfs to ensure the process actually has stopped. Additionally
this change adds a default retry of 5 times with a 2 second sleep
between sending SIGTERM signals to the processes. This retry cound is
dynamic based on the timeout for the stopping command.  If the SIGTERM
does not work, a SIGKILL will also be tried 5 times with a 2 second
sleep before erroring.

Change-Id: I8dcc1c37d17068a9c29480b886d4f1a051f28894
Closes-Bug: 1425579