Update OCF stop actions to use procfs
This change adds a set of fuel functions that can be used to stop
processes using a pid file or a pid. These functions allow for the same
method of stopping processes to be used by all ocf scripts. The process
stopping logic leverages procfs to check to make sure the process has
been stopped.

This change adds a default retry of 5 times with a 2 second sleep
between sending SIGTERM signals to the processes. This retry count is
dynamic based on the timeout for the stopping command. If the SIGTERM
does not work, a SIGKILL will also be tried one time.

This change does not include the ocf script for rabbitmq as that is the
upstream version and won't be leveraging the ocf-fuel-funcs library

Change-Id: I8dcc1c37d17068a9c29480b886d4f1a051f28894
Closes-Bug: 1425579