Move monit module to librarian
This commit removes 'monit' module from fuel-library
and replaces it to be managed via librarian.
This commit uses custom v0.0.4-mos-rc1 tag
which is current master of monit module.
This module can be used as is from upstream master
because all our custom patches are merged in it.

Partial implements: blueprint fuel-puppet-librarian

Change-Id: I78821251bf3874b9c444eead23f3e295a49adb47