Add coverage reports to tests
This change adds simplecov to show the coverage of the noop test files
and also includes the output from the rspec-puppet reports into the
coverage/ folder.

The simplecov report is a report that is merged together based on the
entire noop test run. It will show where we have astute.yaml gaps for
our given noop spec tests.

The rspec-puppet reports are named based on the astute.yaml file that
was used to generate each report.

Change-Id: I9dda550c960fb2742f1f0862305e144a3fb1d9d6
Closes-Bug: #1506557