Create Rakefile to execute lint and spec tasks
This change creates a top level Rakefile that can be used to run all of
the lint, syntax and spec tasks for the modules and files in this
repository.

This is needed to integrate with the openstack-infra puppet syntax and
unit test jobs.

This change disables the unit tests for l23network until they can be
fixed as part of bug 1507742.

Change-Id: I1f637f2f2fd91ce01ba2a92f6ea45499def5333d
Related-Bug: #1507751
Related-Bug: #1507742