Hardcode part_power for swift ring builder
The default part_power for 3 controllers should
be 10. Previous deployments calculated the part_power
of 3 controllers to 10, so this should be preserved
for future deployments.

For custom deployments which require a different part_power,
it should be set in swift_hash in the key 'ring_part_power'.

DocImpact: Should advise users before upgrading from a
prior release about swift ring part_power.
Change-Id: Ic7f8f6824712cc2bb5b530fe1300c9a3ac4ce302
Closes-Bug: #1498368