Fix version override of Fuel release in dockerctl
Corrects options handling did not handle override
of version so that one can correctly use the
following commands:
dockerctl -V 8.0 start postgres
dockerctl --version 7.0 destroy nginx

The wrong variable was in the template leaving the
version defaulted to "", and then reverting to 8.0
as hardocded by the RPM. Switched to @release var
name.

Fixed varname for dockerctl config in hardcoded conf
distributed in the RPM so it does not get replaced.

Change-Id: Ia864b0a9843f927aa85945e5ab0b2da4bec3a440
Related-Commit: I28032852517ec8f7e959ad48a04621ec3e33fea1
Related-Bug: #1501368
Closes-Bug: #1510080