Block all SYN for the HAProxy reload time
And unblock after the reload operation finished.

This introduces a 1 second latency lag to *all* new
TCP connections requests made in the middle of the
HAProxy reload operation (which takes normally ~20 ms).

But otherwise, those of the new connections that targeted
to the HAProxy-backed OpenStack services, would likely
be aborted with RST, which is much worse.

Closes-bug: #1510236

Change-Id: Ic33f7539421fd3a8306e72d2a0b1519d94263380
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>