Block new connections in HAProxy netns for its reload time
And unblock after the reload operation finished.

This introduces a 1 second latency lag to new
TCP connections requests made in the middle of the
HAProxy reload operation (which takes normally ~20 ms).
Otherwise, those would likely be aborted with RST,
which is much worse than a 1 second lag.
Note, that the scope of the blocking rule is limited to the
HAProxy network namespace and doesn't affect other connections.

Closes-bug: #1510236

Change-Id: Ic33f7539421fd3a8306e72d2a0b1519d94263380
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>
Co-author: Sergii Golovatiuk <sgolovatiuk@mirantis.com>