Set swift_partition to /var/lib/glance/node by default
We had an ability to set swift partition in modular
script, but it was removed in this commit:
Ia2428b49a7192a2d7d56ffdc61a75de9ebc816c9

We should use /var/lib/glance/node as swift partition
by default as it was before and provide access
to this folder for swift user.
Also noop tests were added to cover this parameter.

Change-Id: Ic38c726d5ae9fff1815e6a8b4158549c09cec84c
Closes-bug: #1508482