Add retry to unless part of nova flavor m1.micro
The openstack-controller task fails if nova-flavor list returns
a 503 error code, causing the nova flavor-create command to
be run unnecessarily.

Nova flavor-create fails if it already exists, but sometimes
nova flavor-list fails because nova or keystone API is not
ready. The unless statement should retry the same number
of tries until nova flavor-list is able to return a result.

Change-Id: I106dae88fd397c5b3ef137597ad8a4d70e50d134
Closes-Bug: #1511695