Use dockerctl start for supervisorctl container launch
If docker daemon is restarted, the mount points are not safely
torn down. dockerctl cleans these up before trying to start
the container and therefore should be used.

There was an original change to use direct docker commands
because you could not start a container from a different
Fuel release, but now this is possible and works. Therefore,
this patch depends on the patch that fixes it.

Depends-On: Ia864b0a9843f927aa85945e5ab0b2da4bec3a440
Change-Id: Ie3e471d1c394fac9228b7e2bf4b6877775680b3b
Closes-Bug: #1511294