Adjust disk monitoring limits
This change increases the threshold limit for the corosync disk monitor
from 100M to 512M and decreases the disk monitoring interval from 30s to
15s. These changes are to try and provide additional time for pacemaker
to realize a full disk and trigger a failover prior to the disks
actually being filled.

With this change we are also reducing the disk_free_limit configuration
for rabbitmq from the default of 50M to 5M. When rabbitmq hits the
disk_free_limit, it will trigger flow control which will block the
clients. The reduction of the limit from 50M to 5M is to try and still
have the disk limit in place for an emergency situation but provide
additional time for corosync to try and shutdown rabbitmq before it
blocks clients.

DocImpact: Minimum space free set to 512M for a controller. If any
parition on the controller (besides /boot) drops below 512M, the
controller will migrate services off of the affected node. If space
drops below 5M prior to services being migrated, rabbitmq consumers may
be blocked and might have to be restarted before services function
correctly.
Closes-Bug: #1493520

Change-Id: I537543ce04c4f229939fd7836daf1d62c920974b