Add fuel-tasklib to fuel library using tox
This project was deprecated out of stackforge and
needs a home in fuel-library, where it current is used
as a testing library.

Tox is used so it can call tasks-test-requirements.txt
with relative paths correctly. Otherwise, the relative
path would have to be guessed based on where a user
might call it.

Change-Id: I742e167a77cca2ae47436cce7b1baf63e70fe063
Closes-Bug: #1507208
(cherry picked from commit d2cdb1fa973517666d88bd63a7a37586d8e56990)