Update Fuel CI script to use puppet 3.8
This change updates the scripts used by Fuel CI to pull in puppet 3.8
rather than 3.4.

In order to support testing under Puppet 3.8, we have to setup the
puppet settings as part of our overrides setup. This is due to a change
after 3.4 where these settings are no longer setup as part of the
TestHelper inititalize call. This was done as part of PUP-11118.
Without setting up up the settings, rspec may error when we attempt
to access the Puppet scope or functions.

Change-Id: I26c20036cc9a829fc5d7d351530ac358b1a5ace8
Related-Bug: 1500414
Closes-Bug: 1513969