Speedup gathering pids from namespaces by OCF
Currently, neutron OCF scripts are using `lsof` to gather pids of
processes running in particular namespace. `lsof` is extremely slow
due to inoptimal architecture and causes significant timeouts:
2+ minutes on 300 pids in 300 namespaces. On the other hand,
we cannot use `ip netns pids`, as this command is not present in
precise. The fix changes this call to the 2-stage execution inspired
by the original code of `ip netns pids`, which uses comparison of
the namespace inode with the `ns/net` symlink of the running
process that points to the required namespace. This scales about
linear: 1.7 seconds per 300 pids in 300 namespaces.

Closes-Bug: #1464815
Closes-Bug: #1405477

Change-Id: Ife3f5da3558bf4b5ddd0abbe997cd4b845a8d53c