Task for deleting ssh keys and ssl certs generated for deployment
This task will be called during cluster reset to make sure that there
will be new keys generated during next deployment of this env.

validating tasklib checks the graph to be connected, unless the task has
'detached' set to true, in which case it's not checked as part of
deployment graph.

Change-Id: I78203d0922316f02f08955f109c760aa45f8cd8d
Partial-bug: #1507361
Depends-On: I19e253b1ab21da1b556b36cf7d7bfb0b9205581a