Fix supervisorctl interactions for restore
Currently we are always attempting to stop a container in supervisord
without checking that supervisord is actually running. This can lead to
issues where attempting issue a supervisorctl stop while the supervisord
process is stopping. Since we are shutting down the supervisord process
when we do a restore, we can get into a state where the container stop
will actually fail the restore.

This change adds a check to get the supervisord PID to check if it's
running before issuing a supervisorctl stop command. Additionally this
change adds a 60 second sleep after the shutdown of supervisord (as
recommended by supervisord). This sleep will allow the child processes
to exit cleanly and the socket file will be properly cleaned up.

Change-Id: Ic10036109afb26083c60791dd09a71262784e159
Closes-Bug: #1514798