Refactor function configure_default_route
Before the function configure_default_route was a little part of function
generate_network_config which changed default gateway only, due to the provider
for interfaces was not picked correctly.
Now if default route is needed to change we just modify network_scheme
and call generate_network_config with this network_scheme, if no - do
nothing.
Leave only provider lnx as default for l23_stored_config.

Change-Id: I33e88550af5d5cce2886254444ee5d450e578a1c
Closes-bug: #1510072